Add authentication system for mentor form
Create a new table in the database to store single-use authentication
codes. These codes are found in a hidden file, which is read into the
database. Logging into the form requires access to one of these codes,
which is then searched for in the database. If the code is found, the
user is given access to the form.