Make setup.sql idempotent
Running setup.sql throws an error if the database and tables are already
created; change this so that the database and tables are left alone if
they are already created, avoiding errors.

This allows running the script a second time if one or more of the tsv
files was not accessible the first time for whatever reason.