Add a page for mentors to update courses
Fixes #10.
This page is a simplified version of the mentor form, where mentors can
enter classes they're currently taking (or other classes they'd like to
add). This page requires users to sign in with Google, and upon
submission, checks whether the student ID and email match. Then, it adds
new rows to the "connect" table in the database.